TraceRA: outsource SSIBuilder.
